Add ability to stream and pre-fetch results.
We want the ability to tell async fetch results from backend nodes so
that a reader can start pre-fetching results from a backend as it
consumes the queue.

This required using a poller in the client for async i/o support and
allowing a single outstanding RPC at a time.

We also add read_all and write_all abilities to iodevice to fix some
bugs this was causing.