Add in overlapped io method for node in method 1.
Should have tested first, as actually now with a better Record type, the
amount CPU time spent in the heap is so little that overlapping the IO
and heap gains nothing.

Where we need to overlap is the final sort between the linear scan
ending and network transmission starting.

Could either:

1) Start the next linear scan while the sort occurs.
2) Sort in such a way as to end up with the first half of the array
sorted first and available for network transmission.

May want to do both actually.