Add WIP new client side for meth1 that handles multiple-backends better.
Our old client would read chunks of the wire the same as the chunk size
used in the backends for reads. This works poorly as it means the client
runs out of memory very quickly. Now we read from the wire as we need
records (but with a small buffer to minimize syscalls), and when close
to the end of the read, we copy all values of the wire into a larger
buffer so that we have more opportunity for overlapping our next read
with a read at another backend.